Don't do stupid things with primitives; new keywords; -- parsing; ensure modifiers

+++ test/ParserSpec.scala
@@ -74,7 +74,7 @@ class ParserSpec extends FlatSpec with ShouldMatchers {
     val parser = mkParser("""
       class Test {
         void simple();
-        protected int add(int a, long b) { }
+        protected int add(int a, int b) { }
       } """)
     val results = parser.parseClass(NONE).methods
 
@@ -98,7 +98,7 @@ class ParserSpec extends FlatSpec with ShouldMatchers {
 
     val arg_b = add.params(1)
     arg_b.name should be === "b"
-    arg_b.tname.toString should be === "long"
+    arg_b.tname.toString should be === "int"
     arg_b.value should be === None
   }
